requirements.txt
discord.py
peewee
Pillow
overwatch-api
python-decouple
flask
beautifulsoup4 
requests